pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ml-ab-platform"
version = "0.1.0"
description = "Production-quality ML Model A/B Testing Platform with statistical rigor"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "ML A/B Platform" }]
license = { text = "MIT" }
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "pydantic>=2.6",
    "pydantic-settings>=2.2",
    "sqlalchemy>=2.0",
    "scikit-learn>=1.4",
    "xgboost>=2.0",
    "pandas>=2.2",
    "numpy>=1.26",
    "scipy>=1.12",
    "joblib>=1.3",
    "streamlit>=1.31",
    "plotly>=5.19",
    "click>=8.1",
    "structlog>=24.1",
    "pyyaml>=6.0",
    "httpx>=0.27",
    "rich>=13.7",
    "requests>=2.31",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.3",
    "mypy>=1.8",
]

[project.scripts]
mlab = "ml_ab_platform.cli.main:cli"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "A", "C4", "SIM", "RET"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-v --tb=short"
